backend/requirements.txt
fastapi
uvicorn
httpx
supabase
python-dotenv
python-multipart
reportlab
transformers>=4.42
torch
pillow
pymupdf
